6.0

+++ src/components/Nweet.js
@@ -29,19 +29,19 @@ const Nweet = ({ nweetObj, isOwner }) => {
 
   const confirm = useConfirm();
   const onDeleteClick = async () => {
-    confirm({ description: `이 항목을 삭제하실건가요?` })
-      .then(() => console.log("삭제하죠."))
-      .catch(() => console.log("Deletion cancelled."));
+    // confirm({ description: `이 항목을 삭제하실건가요?` })
+    //   .then(() => console.log("삭제하죠."))
+    //   .catch(() => console.log("Deletion cancelled."));
     setNweetModify(null);
-    // const ok = window.confirm("이 Nweet을 삭제 할건가요?");
+    const ok = window.confirm("이 Nweet을 삭제 할건가요?");
 
     // console.log(ok);
-    // if (ok) {
-    await deleteDoc(doc(dbService, `nweets/${nweetObj.id}`), {});
-    if (nweetObj.attachmentUrl) {
-      await deleteObject(ref(storageService, nweetObj.attachmentUrl));
+    if (ok) {
+      await deleteDoc(doc(dbService, `nweets/${nweetObj.id}`), {});
+      if (nweetObj.attachmentUrl) {
+        await deleteObject(ref(storageService, nweetObj.attachmentUrl));
+      }
     }
-    // }
   };
   const toggleEditing = () => setEditing((prev) => !prev);
   const onSubmit = async (event) => {
